wip: improve plot output
This changes from using a tuple of two f64s to
a struct called a PlotDatum. Now it includes
time and total_energy, but next we'll add the
potential and kinetic energy separately, too.
We also now show the x and y values in scientific
form if they are large numbers. That and using
the plotters function to estimate the size of the
text label helps the x and y axes look cleaner
and easier to read.

We also now log DrawingAreaErrors instead of
unwrapping them. Docs for those errors here:
https://docs.rs/plotters/0.3.5/plotters/drawing/enum.DrawingAreaErrorKind.html